feat<sinhvien>: add field malop

+++ app/controllers/sinhvien.php
@@ -5,19 +5,26 @@ class sinhvien extends Controller
   public function index($limit = 5, $offset = 0, $search = "")
   {
     $sinhvienModel = $this->model('sinhvienModel');
-    //$sinhviens = $sinhvienModel->getAllSinhVien();
-    $result = $sinhvienModel->paging($limit, $offset, $search);
-    $sinhviens = $result['sinhviens'];
-    $totalPages = $result['totalPages'];
-    // Trả về View
-    //require_once "../app/views/sinhvien/index.php";
-    $this->view('layout/masterLayout', ['viewname' => 'sinhvien/index', 'sinhviens' => $sinhviens, 'title' => 'Danh sách sinh viên', 'totalPages' => $totalPages]);
+    $result      = $sinhvienModel->paging((int)$limit, (int)$offset, $search);
+    $sinhviens   = $result['sinhviens'];
+    $totalPages  = $result['totalPages'];
+
+    $this->view('layout/masterLayout', [
+      'viewname'   => 'sinhvien/index',
+      'sinhviens'  => $sinhviens,
+      'title'      => 'Danh sách sinh viên',
+      'totalPages' => $totalPages,
+      'offset'     => (int)$offset,
+      'pageSize'   => (int)$limit,
+    ]);
   }
 
   public function create()
   {
+    $lophocModel = $this->model('lophocModel');
+    $lophocs = $lophocModel->getAllLopHoc();
     // Trả về View
-    require_once "../app/views/sinhvien/create.php";
+    $this->view('layout/masterLayout', ['viewname' => 'sinhvien/create', 'title' => 'Tạo sinh viên', 'lophocs' => $lophocs]);
   }
 
   public function store()
@@ -26,9 +33,10 @@ class sinhvien extends Controller
       $MSSV = $_POST['MSSV'];
       $HoTen = $_POST['HoTen'];
       $GioiTinh = $_POST['GioiTinh'];
+      $MaLop = $_POST['MaLop'];
 
       $sinhvienModel = $this->model('sinhvienModel');
-      $result = $sinhvienModel->create($MSSV, $HoTen, $GioiTinh);
+      $result = $sinhvienModel->create($MSSV, $HoTen, $GioiTinh, $MaLop);
       if ($result) {
         header("Location: /sinhvien/index");
         exit();
@@ -44,13 +52,15 @@ class sinhvien extends Controller
     $id = (int)$id;
     $sinhvienModel = $this->model('sinhvienModel');
     $sinhvien = $sinhvienModel->getSinhVienById($id);
+    $lophocModel = $this->model('lophocModel');
+    $lophocs = $lophocModel->getAllLopHoc();
 
     if (!$sinhvien) {
       echo "Sinh viên không tồn tại!";
       exit();
     }
 
-    $this->view('layout/masterLayout', ['viewname' => 'sinhvien/edit', 'sinhvien' => $sinhvien, 'title' => 'Sửa thông tin Sinh viên']);
+    $this->view('layout/masterLayout', ['viewname' => 'sinhvien/edit', 'sinhvien' => $sinhvien, 'title' => 'Sửa thông tin Sinh viên', 'lophocs' => $lophocs]);
   }
 
   public function update($id)
@@ -60,9 +70,9 @@ class sinhvien extends Controller
       $MSSV = $_POST['MSSV'];
       $HoTen = $_POST['HoTen'];
       $GioiTinh = $_POST['GioiTinh'];
-
+      $MaLop = $_POST['MaLop'];
       $sinhvienModel = $this->model('sinhvienModel');
-      $result = $sinhvienModel->update($id, $MSSV, $HoTen, $GioiTinh);
+      $result = $sinhvienModel->update($id, $MSSV, $HoTen, $GioiTinh, $MaLop);
 
       if ($result) {
         header("Location: /sinhvien/index");
@@ -76,15 +86,20 @@ class sinhvien extends Controller
 
   public function delete($id)
   {
-    $id = (int)$id;
-    $sinhvienModel = $this->model('sinhvienModel');
-    $result = $sinhvienModel->delete($id);
+    if (isset($_SERVER['REQUEST_METHOD']) && $_SERVER['REQUEST_METHOD'] === 'POST') {
+      $id = (int)$id;
+      $sinhvienModel = $this->model('sinhvienModel');
+      $result = $sinhvienModel->delete($id);
 
-    if ($result) {
-      header("Location: /sinhvien/index");
-      exit();
+      if ($result) {
+        header("Location: /sinhvien/index");
+        exit();
+      } else {
+        echo "Xoá sinh vien thất bại!";
+        exit();
+      }
     } else {
-      echo "Xoá sinh vien thất bại!";
+      header("Location: /sinhvien/index");
       exit();
     }
   }
